bench: remove libuv dependency
Using time to control the test can lead to non-deterministic results
since the test can run a few cycles more or less.

Calculate the number of tests to run in the specified test window and
replace `libuv` timers with a deterministic loop that starts background
threads to do the memory reads.